<refactor> Refactor BM25
