+++ retrieval/sparse/__init__.py
@@ -1,6 +1,6 @@
 from retrieval.sparse.sparse_base import SparseRetrieval
 from retrieval.sparse.tfidf import TfidfRetrieval
-from retrieval.sparse.bm25 import BM25Retrieval
+from retrieval.sparse.bm25_base import BM25Retrieval
 from retrieval.sparse.atire_bm25 import ATIREBM25Retrieval
 from retrieval.sparse.bm25l import BM25LRetrieval
 from retrieval.sparse.bm25_plus import BM25PlusRetrieval
