+++ retrieval/sparse/bm25_base.py
@@ -1,7 +1,7 @@
-import tqdm
 import pickle
 import numpy as np
 import os.path as p
+from tqdm.auto import tqdm
 
 from konlpy.tag import Mecab
 from transformers import AutoTokenizer
@@ -14,6 +14,11 @@ from utils.tokenization_kobert import KoBertTokenizer
 class BM25Retrieval(SparseRetrieval):
     def __init__(self, args):
         super().__init__(args)
+
+        save_dir = p.join(args.path.embed, self.name)
+        self.encoder_path = p.join(save_dir, f"{self.name}.bin")
+        self.idf_encoder_path = p.join(save_dir, f"{self.name}_idf.bin")
+        self.idf_path = p.join(save_dir, "idf.bin")
 
         if self.args.model.tokenizer_name == "":
             print("Using Mecab tokenizer")
@@ -28,24 +33,38 @@ class BM25Retrieval(SparseRetrieval):
 
         self.b = self.args.retriever.b
         self.k1 = self.args.retriever.k1
-        self.encoder = TfidfVectorizer(tokenizer=self.tokenizer, ngram_range=(1, 2))
+        self.encoder = TfidfVectorizer(tokenizer=self.tokenizer, ngram_range=(1, 2), use_idf=False, norm=None)
+        self.idf_encoder = TfidfVectorizer(tokenizer=self.tokenizer, ngram_range=(1, 2), norm=None, smooth_idf=False)
         self.dls = np.zeros(len(self.contexts))
 
         for idx, context in enumerate(self.contexts):
             self.dls[idx] = len(context)
 
-        self.avdl = sum(self.dls) / len(self.contexts)
+        self.avdl = np.mean(self.dls)
         self.p_embedding = None
+        self.idf = None
 
     def get_embedding(self):
-        if p.isfile(self.embed_path) and p.isfile(self.encoder_path) and not self.args.retriever.retrain:
+        if (
+            p.isfile(self.embed_path)
+            and p.isfile(self.encoder_path)
+            and p.isfile(self.idf_encoder_path)
+            and p.isfile(self.idf_path)
+            and not self.args.retriever.retrain
+        ):
             with open(self.embed_path, "rb") as f:
                 self.p_embedding = pickle.load(f)
 
             with open(self.encoder_path, "rb") as f:
                 self.encoder = pickle.load(f)
+
+            with open(self.idf_encoder_path, "rb") as f:
+                self.idf_encoder = pickle.load(f)
+
+            with open(self.idf_path, "rb") as f:
+                self.idf = pickle.load(f)
         else:
-            self.p_embedding, self.encoder = self._exec_embedding()
+            self.p_embedding, self.encoder, self.idf, self.idf_encoder = self._exec_embedding()
 
             with open(self.embed_path, "wb") as f:
                 pickle.dump(self.p_embedding, f)
@@ -53,40 +72,45 @@ class BM25Retrieval(SparseRetrieval):
             with open(self.encoder_path, "wb") as f:
                 pickle.dump(self.encoder, f)
 
-        self.avdl = self.p_embedding.sum(1).mean()
+            with open(self.idf_path, "wb") as f:
+                pickle.dump(self.idf, f)
+
+            with open(self.idf_encoder_path, "wb") as f:
+                pickle.dump(self.idf_encoder, f)
+
+    def calculate_idf(self):
+        return self.idf_encoder.idf_
 
     def _exec_embedding(self):
-        self.encoder.fit(self.contexts)
-        self.p_embedding = self.encoder.transform(self.contexts)
-        return self.p_embedding, self.encoder
+        self.p_embedding = self.encoder.fit_transform(tqdm(self.contexts, desc="TF calculation: "))
+        self.idf_encoder.fit(tqdm(self.contexts, desc="IDF calculation: "))
+        self.idf = self.calculate_idf()
+
+        return self.p_embedding, self.encoder, self.idf, self.idf_encoder
+
+    def calculate_score(self, p_embedding, query_vec):
+        raise NotImplementedError
 
     def get_relevant_doc_bulk(self, queries, topk):
         query_vecs = self.encoder.transform(queries)
-
-        b, k1, avdl = self.b, self.k1, self.avdl
-        len_p = self.p_embedding.sum(1).A1
 
         doc_scores = []
         doc_indices = []
 
         p_embedding = self.p_embedding.tocsc()
 
-        for query_vec in tqdm.tqdm(query_vecs):
-            p_emb_for_q = p_embedding[:, query_vec.indices]
-            denom = p_emb_for_q + (k1 * (1 - b + b * len_p / avdl))[:, None]
+        self.results = []
 
-            # idf(t) = log [ n / df(t) ] + 1 in sklearn, so it need to be converted
-            # to idf(t) = log [ n / df(t) ] with minus 1
-            idf = self.encoder._tfidf.idf_[None, query_vec.indices] - 1.0
-            numer = p_emb_for_q.multiply(np.broadcast_to(idf, p_emb_for_q.shape)) * (k1 + 1)
+        for query_vec in tqdm(query_vecs):
 
-            result = (numer / denom).sum(1).A1
-
-            if not isinstance(result, np.ndarray):
-                result = result.toarray()
-
+            result = self.calculate_score(p_embedding, query_vec)
+            self.results.append(result)
             sorted_result_idx = np.argsort(result)[::-1]
             doc_score, doc_indice = result[sorted_result_idx].tolist()[:topk], sorted_result_idx.tolist()[:topk]
             doc_scores.append(doc_score)
             doc_indices.append(doc_indice)
+
+        if not isinstance(self.results, np.ndarray):
+            self.results = np.array(self.results)
+
         return doc_scores, doc_indices
